better EPA utilizing proper convex hull expansion
initial proof of concept.  more testing and cleanup necessary.
instead of greedy, chose triangle/plane closest to origin.
terminate when current plane/triangle is determined to be on the minkowski surface.
specific conditions include when hitting the same support vertex again, or when the surface only moves out an insignificant amount.

some references:
 http://movement.stanford.edu/courses/cs468-01-fall/Papers/van-den-bergen.pdf
 https://github.com/melax/sandbox/issues/2
 http://www.gamedev.net/topic/649946-epa-expanding-polytope-algorithm/